pyproject.toml
[project]
name = "unets"
version = "0.1.0"
description = "Implementation of a family of UNet models."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "albumentationsx>=2.0.9",
    "jaxtyping>=0.3.2",
    "lightning>=2.5.2",
    "marimo[recommended]>=0.14.8",
    "torch==2.7.1",
    "torchtyping>=0.1.5",
    "torchvision>=0.22.0",
]

[project.optional-dependencies]
notebooks = ["marimo[recommended]>=0.14.8"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
ignore = ["F722"]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
